toolkit/components/passwordmgr/test/mochitest/test_autocomplete_password_generation_confirm.html

<!DOCTYPE HTML>
<html>
  <head>
  <meta charset="utf-8">
  <title>Test filling generated passwords into confirm password fields</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <script src="/tests/SimpleTest/EventUtils.js"></script>
  <script src="pwmgr_common.js"></script>
  <script src="../../../satchel/test/satchel_common.js"></script>
  <link rel="stylesheet" href="/tests/SimpleTest/test.css" />
</head>
<body>
Login Manager test: filling generated passwords into confirm password fields

<p id="display"></p>
<div id="content"></div>
<pre id="test">
<script class="testbody" type="text/javascript">
  const setupScript = runInParent(function parentTestSetup() {
    const { LoginTestUtils } = ChromeUtils.importESModule(
      "resource://testing-common/LoginTestUtils.sys.mjs"
    );

    addMessageListener(
      "resetLoginsAndGeneratedPasswords", () => {
        LoginTestUtils.clearData();
        LoginTestUtils.resetGeneratedPasswordsCache();
      }
    );
  });

  async function testConfirmPasswordFieldFilledWithGeneratedPassword({
    title,
    html,
    afterFormAutofill,
    expectFilledPasswordConfirmation,
    afterPasswordGeneration
  }) {
    const content = document.querySelector("#content");
    info(`Test confirm password filled with generated password: ${title}`);

    await setupScript.sendAsyncMessage("resetLoginsAndGeneratedPasswords");

    // eslint-disable-next-line no-unsanitized/property
    content.innerHTML = html;

    // wait for the just created form to get autofilled
    await promiseFormsProcessed();

    const form = content.querySelector('form');
    // ensure form did not get initially filled
    is(form.uname.value, "", "Username is not filled");
    is(form.pword.value, "", "Password is not filled");

    if (afterFormAutofill) {
      await afterFormAutofill(form);
    }

    // trigger autocomplete popup
    const { items } = await openPopupOn(form.pword)
    info("triggering password confirmation");

    // examine the popup contents
    checkAutoCompleteResults(items, [
      "Use a Securely Generated Password"
    ], location.host, "Check all rows are correct");

    // Then select generated password
    synthesizeKey("KEY_ArrowDown");
    synthesizeKey("KEY_Enter");

    // and wait until the password has been filled.
    await SimpleTest.promiseWaitForCondition(() => !!form.pword.value, "Wait for generated password to get filled");

    ok(form.pword.matches(":autofill"), "Highlight was applied to the form.pword field");

    const generatedPassword = form.pword.value;
    is(generatedPassword.length, GENERATED_PASSWORD_LENGTH, "Generated password length matches");
    ok(generatedPassword.match(GENERATED_PASSWORD_REGEX), "Generated password format matches");

    LOGIN_FIELD_UTILS.checkPasswordMasked(form.pword, false, "password field is not masked after password generation");

    const passwordInputs = form.querySelectorAll("input[type='password']:not([name='pword'])");
    for (const passwordInput of passwordInputs) {
      LOGIN_FIELD_UTILS.checkPasswordMasked(passwordInput, true, `password field ${passwordInput.name} is masked`);
    }

    ok(!form.uname.matches(":autofill"), "Highlight was not applied to the field");
    is(form.uname.value, "", "Value is still empty")

    if (afterPasswordGeneration) {
      afterPasswordGeneration(form);
    }

    // Check the 2 field values aren't mirrored

    // changing the password field value should result in a message sent to the parent process
    const messageSentPromise = getPasswordEditedMessage();

    // add a character
    form.pword.focus()
    synthesizeKey("KEY_End");
    synthesizeKey("@");

    is(form.pword.value, `${generatedPassword}@`, "Value of the first password field changed");

    // bluring results in a "change" event
    form.pword.blur();

    await messageSentPromise;

    for (const passwordInput of passwordInputs) {
      LOGIN_FIELD_UTILS.checkPasswordMasked(passwordInput, true, `password field ${passwordInput.name} is masked`);
    }

    if (expectFilledPasswordConfirmation === true) {
      is(form.pwordNext.value, generatedPassword, "Value of the confirm field has been filled with generated password");
      ok(form.pwordNext.matches(":autofill"), "Highlight was applied to the input field");
    } else {
      is(form.pwordNext.value, expectFilledPasswordConfirmation, "Value of the confirm field did not change");
      ok(!form.pwordNext.matches(":autofill"), "Highlight was not applied to the input field");
    }

    form.pwordNext.focus();
    form.pwordNext.select();
    synthesizeKey("KEY_Backspace");

    // verify the focused confirm field now masks its input like a normal,
    // non-generated password field after being emptied
    form.pwordNext.focus();
    synthesizeKey("a");
    form.pwordNext.blur();

    for (const passwordInput of passwordInputs) {
      LOGIN_FIELD_UTILS.checkPasswordMasked(passwordInput, true, `password field ${passwordInput.name} is masked`);
    }

    await SimpleTest.promiseWaitForCondition(() => !form.pwordNext.matches(":autofill"), "Highlight was cleared");
  }

  add_task(() => testConfirmPasswordFieldFilledWithGeneratedPassword({
    title: "fill next password field",
    html: `<form id="form1" action="https://example.com" onsubmit="return false;">
    <input type="text" name="uname">
    <input type="password" name="pword" autocomplete="new-password">
    <input type="password" name="pwordNext">
    <button type="submit">Submit</button>
  </form>`,
    expectFilledPasswordConfirmation: true
  }));

  add_task(() => testConfirmPasswordFieldFilledWithGeneratedPassword({
    title: "fill next password field type text",
    html: `<form id="form1" action="https://example.com" onsubmit="return false;">
    <input type="text" name="uname">
    <input type="password" name="pword" autocomplete="new-password">
    <input type="password" name="pwordNext">
    <button type="submit">Submit</button>
  </form>`,
    expectFilledPasswordConfirmation: true,
    afterFormAutofill: form => {
      // manually set type to text. If it's done initially, before promiseFormsProcessed, the field will get a content (in my case a previously filled `a`)
      form.pwordNext.type = "text";
    }
  }));

  add_task(() => testConfirmPasswordFieldFilledWithGeneratedPassword({
    title: "do not fill non empty next password field",
    html: `<form id="form1" action="https://example.com" onsubmit="return false;">
    <input type="text" name="uname">
    <input type="password" name="pword" autocomplete="new-password">
    <input type="password" name="pwordNext" value="initial value">
    <button type="submit">Submit</button>
  </form>`,
    expectFilledPasswordConfirmation: "initial value"
  }));

  add_task(() => testConfirmPasswordFieldFilledWithGeneratedPassword({
    title: "do not fill edited next password field",
    html: `<form id="form1" action="https://example.com" onsubmit="return false;">
    <input type="text" name="uname">
    <input type="password" name="pword" autocomplete="new-password">
    <input type="password" name="pwordNext">
    <button type="submit">Submit</button>
  </form>`,
    expectFilledPasswordConfirmation: "edited value",
    afterFormAutofill: form => {
      form.pwordNext.focus()
      sendString("edited value");
    }
  }));

  add_task(() => testConfirmPasswordFieldFilledWithGeneratedPassword({
    title: "ignore readonly field",
    html: `<form id="form1" action="https://example.com" onsubmit="return false;">
    <input type="text" name="uname">
    <input type="password" name="pword" autocomplete="new-password">
    <input type="password" name="pwordNext" readonly>
    <button type="submit">Submit</button>
  </form>`,
    expectFilledPasswordConfirmation: ""
  }));

  add_task(() => testConfirmPasswordFieldFilledWithGeneratedPassword({
    title: "ignore disabled field",
    html: `<form id="form1" action="https://example.com" onsubmit="return false;">
    <input type="text" name="uname">
    <input type="password" name="pword" autocomplete="new-password">
    <input type="password" name="pwordNext" disabled>
    <button type="submit">Submit</button>
  </form>`,
    expectFilledPasswordConfirmation: ""
  }));

  add_task(() => testConfirmPasswordFieldFilledWithGeneratedPassword({
    title: "prefer matching autocomplete info password field",
    html: `<form id="form1" action="https://example.com" onsubmit="return false;">
    <input type="text" name="uname">
    <input type="password" name="pword" autocomplete="new-password">
    <input type="password" name="pword-between">
    <input type="password" name="pwordNext" autocomplete="new-password">
    <button type="submit">Submit</button>
  </form>`,
    expectFilledPasswordConfirmation: true
  }));

  add_task(() => testConfirmPasswordFieldFilledWithGeneratedPassword({
    title: "ignore disabled matching autocomplete info password field",
    html: `<form id="form1" action="https://example.com" onsubmit="return false;">
    <input type="text" name="uname">
    <input type="password" name="pword" autocomplete="new-password">
    <input type="password" name="pwordNext">
    <input type="password" name="pwordAfter" autocomplete="new-password" disabled>
    <button type="submit">Submit</button>
  </form>`,
    expectFilledPasswordConfirmation: true,
    afterPasswordGeneration: form => {
      is(form.pwordAfter.value, "", "Value of the disabled password did not change");
      ok(!form.pwordAfter.matches(":autofill"), "Highlight was not applied to the disabled password field");
    }
  }));

  add_task(() => testConfirmPasswordFieldFilledWithGeneratedPassword({
    title: "ignore too distant password confirmation field",
    html: `<form id="form1" action="https://example.com" onsubmit="return false;">
    <input type="text" name="uname">
    <input type="password" name="pword" autocomplete="new-password">
    <input type="text" name="junk0">
    <input type="text" name="junk1">
    <input type="text" name="junk2">
    <input type="text" name="junk3">
    <input type="text" name="junk4">
    <input type="password" name="pwordNext" autocomplete="new-password">
    <button type="submit">Submit</button>
  </form>`,
    expectFilledPasswordConfirmation: ""
  }));

  add_task(() => testConfirmPasswordFieldFilledWithGeneratedPassword({
    title: "too many disabled fields",
    html: `<form id="form1" action="https://example.com" onsubmit="return false;">
    <input type="text" name="uname">
    <input type="password" name="pword" autocomplete="new-password">
    <input type="text" name="junk0" disabled>
    <input type="text" name="junk1" disabled>
    <input type="text" name="junk2" disabled>
    <input type="text" name="junk3" disabled>
    <input type="text" name="junk4" disabled>
    <input type="password" name="pwordNext" autocomplete="new-password">
    <button type="submit">Submit</button>
  </form>`,
    expectFilledPasswordConfirmation: ""
  }));

  add_task(() => testConfirmPasswordFieldFilledWithGeneratedPassword({
    title: "skip over hidden fields",
    html: `<form id="form1" action="https://example.com" onsubmit="return false;">
    <input type="text" name="uname">
    <input type="password" name="pword" autocomplete="new-password">
    <input type="hidden" name="junk0">
    <input type="hidden" name="junk1">
    <input type="hidden" name="junk2">
    <input type="hidden" name="junk3">
    <input type="hidden" name="junk4">
    <input type="password" name="pwordNext" autocomplete="new-password">
    <button type="submit">Submit</button>
  </form>`,
    expectFilledPasswordConfirmation: true
  }));

  // this form has a password field after the confirm-password field we don't want to fill into
  add_task(() => testConfirmPasswordFieldFilledWithGeneratedPassword({
    title: "do not fill third password field",
    html: `<form id="form1" action="https://example.com" onsubmit="return false;">
    <input type="text" name="uname">
    <input type="password" name="pword" autocomplete="new-password">
    <input type="password" name="pwordNext" autocomplete="new-password">
    <input type="password" name="pwordExtra" autocomplete="new-password">
    <button type="submit">Submit</button>
  </form>`,
    expectFilledPasswordConfirmation: true,
    afterPasswordGeneration: form => {
      is(form.pwordExtra.value, "", "Value of the extrafield did not change");
      ok(!form.pwordExtra.matches(":autofill"), "Highlight was not applied to the extra field");
    }
  }));
</script>
</pre>
</body>
</html>
